+++ imports.py
@@ -27,7 +27,7 @@ from kivy.uix.anchorlayout import AnchorLayout
 from kivy.uix.label import Label
 from kivy.core.text import LabelBase
 from kivy.uix.checkbox import CheckBox
-from kivy.uix.button import Button
+from kivy.uix.button import Button, ButtonBehavior
 from kivy.properties import ObjectProperty
 from kivy.core.window import Window
 from kivy.uix.behaviors import ToggleButtonBehavior